% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_transition_risk_category.R
\name{add_transition_risk_category}
\alias{add_transition_risk_category}
\title{Adds transition risk categories for transition risk scores}
\usage{
add_transition_risk_category(data)
}
\arguments{
\item{data}{A dataframe}
}
\value{
A dataframe.
}
\description{
Adds transition risk categories for transition risk scores
}
\examples{
library(tiltToyData)
library(readr)
library(dplyr)
options(readr.show_col_types = FALSE)

emissions_profile_products <- read_csv(toy_emissions_profile_products_ecoinvent())
all_uuids_scenario_sectors <- read_csv(toy_all_uuids_scenario_sectors())
scenarios <- read_csv(toy_sector_profile_any_scenarios())

transition_risk_thresholds <- add_thresholds_transition_risk(
  emissions_profile_products,
  all_uuids_scenario_sectors,
  scenarios
)

output <- add_transition_risk_category(transition_risk_thresholds)
output
}
\keyword{internal}
