% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_case_worst_case_transition_risk_profile.R
\name{best_case_worst_case_transition_risk_profile}
\alias{best_case_worst_case_transition_risk_profile}
\title{Calculates best case and worst case for tranistion risk profile at product level}
\usage{
best_case_worst_case_transition_risk_profile(data)
}
\arguments{
\item{data}{Dataframe. Transition risk profile product level output}
}
\value{
A dataframe
}
\description{
Calculates best case and worst case for tranistion risk profile at product level
}
\keyword{internal}
