% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data_tr_score.R
\name{example_emissions_profile_at_product_level}
\alias{example_emissions_profile_at_product_level}
\alias{example_sector_profile_at_product_level}
\title{Example input datasets for Transition Risk Score}
\usage{
example_emissions_profile_at_product_level()

example_sector_profile_at_product_level()
}
\value{
A dataframe.
}
\description{
Example input datasets for Transition Risk Score
}
\examples{
example_emissions_profile_at_product_level()
example_sector_profile_at_product_level()
}
\keyword{internal}
