% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_emissions_profile_trs}
\alias{example_emissions_profile_trs}
\alias{example_sector_profile_trs}
\title{Example input datasets for Transition Risk Score}
\usage{
example_emissions_profile_trs()

example_sector_profile_trs()
}
\value{
A dataframe.
}
\description{
Example input datasets for Transition Risk Score
}
\examples{
example_emissions_profile_trs()
example_sector_profile_trs()
}
\keyword{internal}
