% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_wider_transition_risk_profile.R
\name{pivot_wider_transition_risk_profile}
\alias{pivot_wider_transition_risk_profile}
\title{Pivot company-level columns to wide format for indicator "transition risk profile"}
\usage{
pivot_wider_transition_risk_profile(data, include_co2 = FALSE)
}
\arguments{
\item{include_co2}{Logical. Include \verb{co2_*} columns ?}
}
\value{
A Dataframe
}
\description{
Pivot company-level columns to wide format for indicator "transition risk profile"
}
\examples{
\dontrun{
library(readr, warn.conflicts = FALSE)
library(dplyr, warn.conflicts = FALSE)
library(tiltToyData, warn.conflicts = FALSE)
library(tiltIndicator)
library(tiltIndicatorAfter)

restore <- options(list(
  readr.show_col_types = FALSE,
  tiltIndicatorAfter.output_co2_footprint = TRUE
))

toy_emissions_profile_products_ecoinvent <- read_csv(toy_emissions_profile_products_ecoinvent())
toy_emissions_profile_any_companies <- read_csv(toy_emissions_profile_any_companies())
toy_sector_profile_any_scenarios <- read_csv(toy_sector_profile_any_scenarios())
toy_sector_profile_companies <- read_csv(toy_sector_profile_companies())
toy_europages_companies <- read_csv(toy_europages_companies())
toy_ecoinvent_activities <- read_csv(toy_ecoinvent_activities())
toy_ecoinvent_europages <- read_csv(toy_ecoinvent_europages())
toy_ecoinvent_inputs <- read_csv(toy_ecoinvent_inputs())
toy_isic_name <- read_csv(toy_isic_name())
toy_all_activities_scenario_sectors <- read_csv(toy_all_activities_scenario_sectors())

toy_emissions_profile <- profile_emissions(
  companies = toy_emissions_profile_any_companies,
  co2 = toy_emissions_profile_products_ecoinvent,
  europages_companies = toy_europages_companies,
  ecoinvent_activities = toy_ecoinvent_activities,
  ecoinvent_europages = toy_ecoinvent_europages,
  isic = toy_isic_name
)

toy_sector_profile <- profile_sector(
  companies = toy_sector_profile_companies,
  scenarios = toy_sector_profile_any_scenarios,
  europages_companies = toy_europages_companies,
  ecoinvent_activities = toy_ecoinvent_activities,
  ecoinvent_europages = toy_ecoinvent_europages,
  isic = toy_isic_name
)

wide_format <- transition_risk_profile_impl(
  emissions_profile,
  sector_profile,
  co2,
  all_activities_scenario_sectors,
  scenarios
) |>
  add_transition_risk_category_at_company_level() |>
  best_case_worst_case_transition_risk_profile_at_company_level() |>
  pivot_wider_transition_risk_profile(include_co2 = TRUE) |>
  unnest_company()
wide_format

# Cleanup
options(restore)
}
}
\keyword{internal}
