% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_transition_risk.R
\name{profile_transition_risk}
\alias{profile_transition_risk}
\title{Calculate the indicator "transition risk profile"}
\usage{
profile_transition_risk(
  emissions_profile,
  sector_profile,
  pivot_wider = FALSE,
  emissions_profile_products,
  all_uuids_scenario_sectors,
  scenarios
)
}
\arguments{
\item{emissions_profile}{Nested data frame. The output of
\code{profile_emissions()}.}

\item{sector_profile}{Nested data frame. The output of \code{profile_sector()}.}

\item{pivot_wider}{Logical. Pivot the output at company level to a wide
format?}

\item{emissions_profile_products}{A dataframe}

\item{all_uuids_scenario_sectors}{A dataframe}

\item{scenarios}{A dataframe}
}
\value{
A data frame with the column \code{companies_id}, and the nested
columns\code{product} and \code{company} holding the outputs at product and company
level.
}
\description{
Adds the risk classification to calculated transition risk scores from
emission profile and sector profile indicator.
}
\examples{
library(readr, warn.conflicts = FALSE)
library(dplyr, warn.conflicts = FALSE)
library(tiltToyData, warn.conflicts = FALSE)
library(tiltIndicatorAfter, warn.conflicts = FALSE)

set.seed(123)
restore <- options(list(
  readr.show_col_types = FALSE,
  tiltIndicatorAfter.output_co2_footprint = TRUE
))

toy_emissions_profile_products_ecoinvent <- read_csv(toy_emissions_profile_products_ecoinvent())
toy_emissions_profile_any_companies <- read_csv(toy_emissions_profile_any_companies())
toy_sector_profile_any_scenarios <- read_csv(toy_sector_profile_any_scenarios())
toy_sector_profile_companies <- read_csv(toy_sector_profile_companies())
toy_europages_companies <- read_csv(toy_europages_companies())
toy_ecoinvent_activities <- read_csv(toy_ecoinvent_activities())
toy_ecoinvent_europages <- read_csv(toy_ecoinvent_europages())
toy_ecoinvent_inputs <- read_csv(toy_ecoinvent_inputs())
toy_isic_name <- read_csv(toy_isic_name())

emissions_profile <- profile_emissions(
  companies = toy_emissions_profile_any_companies,
  co2 = toy_emissions_profile_products_ecoinvent,
  europages_companies = toy_europages_companies,
  ecoinvent_activities = toy_ecoinvent_activities,
  ecoinvent_europages = toy_ecoinvent_europages,
  isic = toy_isic_name
)

sector_profile <- profile_sector(
  companies = toy_sector_profile_companies,
  scenarios = toy_sector_profile_any_scenarios,
  europages_companies = toy_europages_companies,
  ecoinvent_activities = toy_ecoinvent_activities,
  ecoinvent_europages = toy_ecoinvent_europages,
  isic = toy_isic_name
)

emissions_profile_products <- read_csv(toy_emissions_profile_products_ecoinvent())
all_uuids_scenario_sectors <- read_csv(toy_all_uuids_scenario_sectors())
scenarios <- read_csv(toy_sector_profile_any_scenarios())

output <- profile_transition_risk(emissions_profile,
  sector_profile,
  pivot_wider = FALSE,
  emissions_profile_products,
  all_uuids_scenario_sectors,
  scenarios
)

output |> unnest_product()

output |> unnest_company()
}
