% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiltTransitionRisk-options.R
\name{tiltIndicatorAfter_options}
\alias{tiltIndicatorAfter_options}
\title{tiltTransitionRisk options}
\description{
These options are meant to be used mainly by developers or analysts while
testing the code or creating data:
\itemize{
\item \code{tiltIndicatorAfter.output_co2_footprint}:
\itemize{
\item At product level it outputs licensed column \code{co2_footprint}.
\item At company level it outputs the column \code{co2_avg} (average \code{co2_footprint}
by \code{companies_id}).
}
\item \code{tiltIndicatorAfter.verbose}: Controls verbosity.
}
}
\examples{
library(readr, warn.conflicts = FALSE)
library(dplyr, warn.conflicts = FALSE)
library(withr)
library(tiltToyData)
library(tiltIndicator)
library(tiltIndicatorAfter)

set.seed(1)

restore <- options(list(
  readr.show_col_types = FALSE,
  tiltIndicatorAfter.verbose = TRUE,
  tiltIndicatorAfter.output_co2_footprint = TRUE
))

companies <- read_csv(toy_emissions_profile_any_companies())
products <- read_csv(toy_emissions_profile_products_ecoinvent())
europages_companies <- read_csv(toy_europages_companies())
ecoinvent_activities <- read_csv(toy_ecoinvent_activities())
ecoinvent_europages <- read_csv(toy_ecoinvent_europages())
isic_name <- read_csv(toy_isic_name())

result <- profile_emissions(
  companies,
  products,
  europages_companies = europages_companies,
  ecoinvent_activities = ecoinvent_activities,
  ecoinvent_europages = ecoinvent_europages,
  isic = isic_name
)

result |>
  unnest_product() |>
  select(matches(c("co2")))

result |>
  unnest_company() |>
  select(matches(c("co2")))
}
\keyword{internal}
