% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_all_uuids_scenario_sectors.R
\name{toy_all_uuids_scenario_sectors}
\alias{toy_all_uuids_scenario_sectors}
\title{Adds toy data for \code{all_uuids_scenario_sectors}}
\usage{
toy_all_uuids_scenario_sectors()
}
\description{
Adds toy data for \code{all_uuids_scenario_sectors}
}
\keyword{internal}
