% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_risk_profile.R
\name{transition_risk_profile}
\alias{transition_risk_profile}
\title{DRAFT of transition risk}
\usage{
transition_risk_profile(
  products,
  products_scenarios,
  scenarios,
  low_threshold = tercile(.data$transition_risk_score, 1),
  high_threshold = tercile(.data$transition_risk_score, 2)
)
}
\arguments{
\item{products, products_scenarios, scenarios}{TODO Data frames.}

\item{low_threshold, high_threshold}{TODO Expressions.}
}
\value{
TODO A data frame
}
\description{
DRAFT of transition risk
}
\examples{
library(tiltToyData)
library(readr, warn.conflicts = FALSE)
library(dplyr, warn.conflicts = FALSE)

restore <- options(readr.show_col_types = FALSE)

products <- read_csv(toy_emissions_profile_products_ecoinvent())
# TODO: Submit a PR to tiltToyData?
products_scenarios <- read_csv(toy_sector_profile_companies()) |>
  select(-c("companies_id", "company_name", "clustered")) |>
  distinct()
scenarios <- read_csv(toy_sector_profile_any_scenarios())

transition_risk_profile(products, products_scenarios, scenarios)

options(restore)
}
