% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_thresholds_transition_risk.R
\name{add_benchmark_tr_score}
\alias{add_benchmark_tr_score}
\title{Calulate \code{benchmark_tr_score} column}
\usage{
add_benchmark_tr_score(
  data,
  col_ranking = col_ranking(),
  col_target = col_targets()
)
}
\arguments{
\item{data}{Dataframe.}

\item{profile_ranking}{Dataframe column.}

\item{reduction_targets}{Dataframe column.}
}
\description{
Calulate \code{benchmark_tr_score} column
}
\keyword{internal}
