% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiltTransitionRisk-package.R
\docType{package}
\name{tiltTransitionRisk-package}
\alias{tiltTransitionRisk}
\alias{tiltTransitionRisk-package}
\title{tiltTransitionRisk: Transition Risk}
\description{
Transition risk.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/2DegreesInvesting/tiltTransitionRisk}
  \item \url{https://2degreesinvesting.github.io/tiltTransitionRisk/}
  \item Report bugs at \url{https://github.com/2DegreesInvesting/tiltTransitionRisk/issues}
}

}
\keyword{internal}
